Rough draft of comprehensive test
The project tries to use CodeExport in as many different ways as
possible while remaining simple. I admit, some of the code is strange,
but it is not hard to understand. I haven't used the same code style
as the rest of the code. I will fix this up in the next commit.